requirements.txt
click>=8.0
PyYAML>=6.0
feedparser>=6.0
requests>=2.0
openai>=0.27.0
GitPython>=3.1
python-dateutil>=2.8
urllib3<2.0
python-dotenv>=0.21.0
